requirements.txt
asgiref==1.1.2
attrs==17.3.0
autobahn==17.10.1
Automat==0.6.0
backports.functools-lru-cache==1.5
backports.shutil-get-terminal-size==1.0.0
boto==2.48.0
boto3==1.4.7
botocore==1.7.43
certifi==2017.11.5
channels==1.1.8
chardet==3.0.4
configparser==3.5.0
constantly==15.1.0
cssselect==1.0.3
cycler==0.10.0
daphne==1.3.0
Django==1.11.7
django-bootstrap-breadcrumbs==0.8.2
django-datetime-widget==0.9.3
django-sendgrid-v5==0.6.84
django-storages==1.6.5
django-widget-tweaks==1.4.1
docutils==0.14
enum34==1.1.6
fake-useragent==0.1.10
flake8==3.5.0
futures==3.1.1
hyperlink==17.3.1
idna==2.6
incremental==17.5.0
jmespath==0.9.3
kiwisolver==1.0.1
lxml==4.2.2
matplotlib==2.2.3
mccabe==0.6.1
numpy==1.14.0
olefile==0.44
opencv-python==3.2.0.6
pandas==0.22.0
parse==1.8.4
pathlib==1.0.1
pew==1.1.0
Pillow==4.3.0
pipenv==8.3.2
poster==0.8.1
psycopg2==2.7.3.1
pycodestyle==2.3.1
pyflakes==1.6.0
pyparsing==2.3.0
pyquery==1.4.0
python-dateutil==2.6.1
python-http-client==3.0.0
pytz==2017.3
requests==2.18.4
requests-html==0.2.2
s3transfer==0.1.11
sendgrid==5.3.0
shutilwhich==1.1.0
six==1.11.0
subprocess32==3.5.3
txaio==2.8.2
Unirest==1.1.7
urllib3==1.22
virtualenv==15.1.0
virtualenv-clone==0.2.6
whitenoise==3.3.1
zope.interface==4.4.3
